refactor(Lotto): 상수 병합
