--- src/main/java/lotto/common/constant/Lotto.java
@@ -1,11 +1,0 @@
-package lotto.common.constant;
-
-/**
- * @author : jiffyin7@gmail.com
- * @since : 24. 11. 1.
- */
-public class Lotto {
-  public static final int LOTTO_MINIMUM_NUMBER = 1;
-  public static final int LOTTO_MAXIMUM_NUMBER = 45;
-  public static final int LOTTO_NUMBER_COUNT = 6;
-}
